Adds two simple features to the SSHOperator
First, adds a named parameter to the SSHOperator's constructor for ssh_conn_id (and makes either
ssh_hook or ssh_conn_id required). Second, adds a named parameter to the SSHOperator's constructor
for xcom_push_type, defaulting to line, and allows the full results of the command to be returned as a list of lines.